lsasm unified & localized.
