+++ Resources/Locale/en-US/commands.ftl
@@ -71,3 +71,8 @@ cmd-mem-help = Usage: mem
 
 cmd-mem-report = Heap Size: { TOSTRING($heapSize, "N0") }
     Total Allocated: { TOSTRING($totalAllocated, "N0") }
+
+## 'lsasm' command
+cmd-lsasm-desc = Lists loaded assemblies by load context
+cmd-lsasm-help = Usage: lsasm
+
